cheat

+++ app/Http/Controllers/Api/V1/RoleController.php
@@ -14,6 +14,7 @@ use App\Traits\ApiResponse;
 use App\Traits\ValidationResponse;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Response;
 
 class RoleController extends Controller
@@ -29,11 +30,17 @@ class RoleController extends Controller
 
     public function index()
     {
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         return new RoleCollection($this->roleService->getAllRoles());
     }
 
     public function store(RoleRequest $request)
     {
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         if ($this->roleService->roleExists($request->input('name'), $request->input('persian_name'))) {
             return $this->failedValidationResponse('این نقش قبلا ثبت شده است');
         }
@@ -50,7 +57,9 @@ class RoleController extends Controller
 //                'error' => 'این نقش قبلا ثبت شده است'
 //            ]);
 //        }
-
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         $role = $this->roleService->updateRole($request, $role);
         if ($role) {
             return $this->apiResponse(null);
@@ -61,6 +70,9 @@ class RoleController extends Controller
 
     public function changeStatus(Role $role, $status)
     {
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         $updated = $this->roleService->updateRoleStatus($role, $status);
         if ($updated) {
             return $this->apiResponse(null);
@@ -71,6 +83,9 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         $roleDelete = $this->roleService->deleteRole($role->id);
         $result = (bool)$roleDelete;
         return $this->apiResponse(null, hasError: !$result);
@@ -78,12 +93,18 @@ class RoleController extends Controller
 
     public function showRolePermissions(Role $role)
     {
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         $rolePermissions = $this->roleService->getRolePermissions($role);
         return new PermissionCollection($rolePermissions);
     }
 
     public function storeRolePermissions(Request $request, Role $role)
     {
+        if (Gate::denies('admin')) {
+            return $this->apiResponse(null, 403, 'forbidden', true);
+        }
         try {
             $this->roleService->addPermissionToRole($request, $role);
             return $this->apiResponse(null);
